refactor(electron): clarify sidecar packaging helpers
