+++ apps/electron/scripts/prepare-electron-sidecars.ts
@@ -1,13 +1,20 @@
-import { spawn } from "node:child_process";
-import { createHash } from "node:crypto";
 import type { Stats } from "node:fs";
-import { chmod, copyFile, mkdir, readFile, rm, stat, writeFile } from "node:fs/promises";
+import { chmod, copyFile, mkdir, rm, stat } from "node:fs/promises";
 import { dirname, join, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
 import { $ } from "bun";
 import {
+  archiveEntryPathToFilePath,
+  type DownloadElectronSidecarAsset,
+  downloadElectronSidecarAsset,
+  type ExtractElectronSidecarArchive,
+  extractElectronSidecarArchive,
+  prepareCachedElectronSidecarArchive,
+  type VerifyElectronSidecarArchiveChecksum,
+  verifyElectronSidecarArchiveChecksum,
+} from "./electron-sidecar-archives";
+import {
   ELECTRON_SIDECAR_IDS,
-  type ElectronExternalSidecarAsset,
   type ElectronSidecarId,
   electronExternalSidecarAssetFileName,
   electronSidecarDisplayName,
@@ -34,27 +41,13 @@ type ResolveElectronSidecarBuildPlanInput = {
   workspaceRoot: string;
 };
 
-type DownloadSidecarAssetInput = {
-  asset: ElectronExternalSidecarAsset;
-  archivePath: string;
-};
-
-type ExtractSidecarArchiveInput = {
-  archivePath: string;
-  asset: ElectronExternalSidecarAsset;
-  extractionDirectory: string;
-};
-
 type PrepareElectronSidecarsInput = ResolveElectronSidecarBuildPlanInput & {
   arch: ElectronReleaseArch;
   chmodFile?: (path: string, mode: number) => Promise<void>;
   compileMcp?: (plan: ElectronSidecarBuildPlan) => Promise<void>;
-  downloadAsset?: (input: DownloadSidecarAssetInput) => Promise<void>;
-  extractArchive?: (input: ExtractSidecarArchiveInput) => Promise<void>;
-  verifyArchiveChecksum?: (
-    asset: ElectronExternalSidecarAsset,
-    archivePath: string,
-  ) => Promise<void>;
+  downloadAsset?: DownloadElectronSidecarAsset;
+  extractArchive?: ExtractElectronSidecarArchive;
+  verifyArchiveChecksum?: VerifyElectronSidecarArchiveChecksum;
 };
 
 const toElectronReleasePlatform = (platform: NodeJS.Platform): ElectronReleasePlatform => {
@@ -68,9 +61,6 @@ const toElectronReleaseArch = (arch: NodeJS.Architecture): ElectronReleaseArch =
   if (arch === "arm64" || arch === "x64") return arch;
   throw new Error(`Unsupported Electron sidecar host architecture: ${arch}`);
 };
-
-const archiveEntryPathToFilePath = (entryPath: string): string[] =>
-  entryPath.replace(/\\/g, "/").replace(/^\.\//, "").split("/").filter(Boolean);
 
 export const resolveElectronSidecarBuildPlan = ({
   electronPackageDirectory,
@@ -92,27 +82,6 @@ export const resolveElectronSidecarBuildPlan = ({
     workspaceRoot,
   };
 };
-
-const runCommand = (command: string, args: string[], cwd: string): Promise<void> =>
-  new Promise((resolveCommand, rejectCommand) => {
-    const child = spawn(command, args, {
-      cwd,
-      shell: process.platform === "win32",
-      stdio: "inherit",
-    });
-
-    child.on("error", rejectCommand);
-    child.on("exit", (exitCode) => {
-      if (exitCode === 0) {
-        resolveCommand();
-        return;
-      }
-
-      rejectCommand(
-        new Error(`${command} ${args.join(" ")} exited with code ${exitCode ?? "unknown"}`),
-      );
-    });
-  });
 
 const assertSidecarFile = async ({
   label,
@@ -154,91 +123,6 @@ const assertFileExists = async (path: string, label: string): Promise<void> => {
   }
 
   throw new Error(`${label} is missing: ${path}`);
-};
-
-const sha256File = async (path: string): Promise<string> =>
-  createHash("sha256")
-    .update(await readFile(path))
-    .digest("hex");
-
-const verifyArchiveChecksum = async (
-  asset: ElectronExternalSidecarAsset,
-  archivePath: string,
-): Promise<void> => {
-  const actualSha256 = await sha256File(archivePath);
-  if (actualSha256 !== asset.sha256) {
-    throw new Error(
-      `Checksum mismatch for Electron ${electronSidecarDisplayName(asset.id)} sidecar asset ${electronExternalSidecarAssetFileName(
-        asset,
-      )}: expected ${asset.sha256}, got ${actualSha256}.`,
-    );
-  }
-};
-
-const downloadAsset = async ({ archivePath, asset }: DownloadSidecarAssetInput): Promise<void> => {
-  const response = await fetch(asset.url, {
-    headers: { "User-Agent": "openducktor-electron-sidecar-builder" },
-  });
-  if (!response.ok) {
-    throw new Error(
-      `Failed downloading Electron ${electronSidecarDisplayName(asset.id)} sidecar asset from ${
-        asset.url
-      }: HTTP ${response.status}.`,
-    );
-  }
-
-  const bytes = Buffer.from(await response.arrayBuffer());
-  if (bytes.length === 0) {
-    throw new Error(
-      `Downloaded Electron ${electronSidecarDisplayName(asset.id)} sidecar asset is empty: ${
-        asset.url
-      }`,
-    );
-  }
-
-  await mkdir(dirname(archivePath), { recursive: true });
-  await writeFile(archivePath, bytes);
-};
-
-const prepareCachedArchive = async ({
-  archivePath,
-  asset,
-  download,
-  verifyChecksum,
-}: DownloadSidecarAssetInput & {
-  download: (input: DownloadSidecarAssetInput) => Promise<void>;
-  verifyChecksum: (asset: ElectronExternalSidecarAsset, archivePath: string) => Promise<void>;
-}): Promise<void> => {
-  try {
-    const metadata = await stat(archivePath);
-    if (!metadata.isFile()) {
-      throw new Error("expected a file but found a non-file entry");
-    }
-    if (metadata.size === 0) {
-      throw new Error("expected a non-empty file");
-    }
-    await verifyChecksum(asset, archivePath);
-    return;
-  } catch (cause) {
-    if ((cause as NodeJS.ErrnoException).code !== "ENOENT") {
-      throw cause;
-    }
-  }
-
-  await download({ archivePath, asset });
-  await verifyChecksum(asset, archivePath);
-};
-
-const extractArchive = async ({
-  archivePath,
-  asset,
-  extractionDirectory,
-}: ExtractSidecarArchiveInput): Promise<void> => {
-  const extractArgs =
-    asset.archiveType === "tar.gz"
-      ? ["-xzf", archivePath, "-C", extractionDirectory, asset.executablePath]
-      : ["-xf", archivePath, "-C", extractionDirectory, asset.executablePath];
-  await runCommand("tar", extractArgs, extractionDirectory);
 };
 
 const compileMcpSidecar = async (plan: ElectronSidecarBuildPlan): Promise<void> => {
@@ -291,13 +175,13 @@ const stageExternalSidecar = async ({
 }: {
   arch: ElectronReleaseArch;
   chmodFile: (path: string, mode: number) => Promise<void>;
-  download: (input: DownloadSidecarAssetInput) => Promise<void>;
+  download: DownloadElectronSidecarAsset;
   electronPackageDirectory: string;
-  extract: (input: ExtractSidecarArchiveInput) => Promise<void>;
+  extract: ExtractElectronSidecarArchive;
   plan: ElectronSidecarBuildPlan;
   platform: ElectronReleasePlatform;
   sidecarId: "beads" | "dolt";
-  verifyChecksum: (asset: ElectronExternalSidecarAsset, archivePath: string) => Promise<void>;
+  verifyChecksum: VerifyElectronSidecarArchiveChecksum;
 }): Promise<PreparedElectronSidecar> => {
   const asset = resolveElectronExternalSidecarAsset({ arch, id: sidecarId, platform });
   const archivePath = join(
@@ -315,7 +199,7 @@ const stageExternalSidecar = async ({
     `${sidecarId}-${platform}-${arch}`,
   );
 
-  await prepareCachedArchive({ archivePath, asset, download, verifyChecksum });
+  await prepareCachedElectronSidecarArchive({ archivePath, asset, download, verifyChecksum });
   await rm(extractionDirectory, { force: true, recursive: true });
   await mkdir(extractionDirectory, { recursive: true });
   await extract({ archivePath, asset, extractionDirectory });
@@ -345,9 +229,9 @@ export const prepareElectronSidecars = async ({
   arch,
   chmodFile = chmod,
   compileMcp = compileMcpSidecar,
-  downloadAsset: download = downloadAsset,
-  extractArchive: extract = extractArchive,
-  verifyArchiveChecksum: verifyChecksum = verifyArchiveChecksum,
+  downloadAsset: download = downloadElectronSidecarAsset,
+  extractArchive: extract = extractElectronSidecarArchive,
+  verifyArchiveChecksum: verifyChecksum = verifyElectronSidecarArchiveChecksum,
   ...input
 }: PrepareElectronSidecarsInput): Promise<{
   plan: ElectronSidecarBuildPlan;
